Change order metadata types (gnosis/gp-v2-services#1696)
This PR just makes a small change to the order MetaData type.

Specifically, the executed sell amount before fees and executed fee amounts are limited to fit within a `U256` because they are capped at amounts specified in the signed order (the `sell_amount` parameter - an exact amount for sell orders and upper bound for buy orders - and the `fee_amount` parameter).

The motivation for this small change is for these values to already be `U256` to help make code for computation around partially fillable orders simpler.

### Test Plan

Rust's type system and compiler. Also adjusted existing unit test to show all possibilities of `executed_*_amount` overflows.