Add Swapr-specific `PoolReading` Implementation (gnosis/gp-v2-services#1720)
Follow up to #1719

This PR adds a Swapr-specific `PoolReading` implementation that wraps the exising `DefaultPoolReader` to additionally query AMM fees. This is required because:
- The fee defaults to something different than Uniswap (0.25% instead of 0.3%)
- Can be modified at any point in time by some DEX admin (so can't be read once and cached forever).

Note that this reader is still not being used. In order to hook it up to the current `PoolAggregator` that fetches all Uniswap-v2-like pools requires some additional refactoring that will come in a follow up PR.

### Test Plan

Added some unit tests around the logic for building a pool for its asynchronously fetched parts. Also added a manual ignore test:
```
% cargo test -p shared -- --ignored --nocapture fetch_swapr_pool
WETH <> wxDAI pool: Pool {
    tokens: TokenPair(
        0x6a023ccd1ff6f2045c3309768ead9e68f978f6e1,
        0xe91d153e0b41518a2ce8dd3d7944fa863463a97d,
    ),
    reserves: (
        379003626006978207230,
        979677523238372752737977,
    ),
    fee: Ratio {
        numer: 1,
        denom: 400,
    },
}
```

Notice that the fee is `1/400 = 0.0025 = 0.25% != 0.3%` 🎉.