Support encoding approvals for HTTP solvers (gnosis/gp-v2-services#1746)
This PR adds support for encoding ERC20 approvals on behalf of solvers.

The motivation behind this change is that the HTTP solver will get liquidity from sources other than the driver (Uniswap V3 and Curve for example). Because of this, it will now send settlements with raw interactions and will it will have to encode approvals to the respective token transfer proxy contracts (for example the Uniswap V3 router).

This change adds a convenient `approvals` field to the solution model that the HTTP solvers return. Approvals specified this way use the common logic to determine if an approval interaction is actually needed or if the settlement contract already has sufficient allowance.

See #1744 for more details.

### Test Plan

Added new unit test to verify approval folding logic and another to check deserialization works as expected.