Additional settlement logging (#213)
One small step towards #209

This PR adds some additional logging around received solutions. Specifically it will:
- Log **all** settlements from any solver before filtering
- Log settlements that are considered for the settlement competition (pre-simulation)
- Adds more context for why HTTP responses from solvers may be ignored

This is a kind of _"low effort"_ additional logging as it does not always specifically give reasons as to why specific solutions are being filtered (we just count if any solutions are being filtered per solver) and is a little brittle (boiler plate for each additional "filter" we add, doesn't solve a bit of "spaghetti" from filtering settlements at various point in the run loop). Still it is an improvement over what we have now since external solvers provide a single settlement - so we can tell for sure why an external solver settlement is being ignored.