Allow interactions to be internal (#195)
Now that we have external solvers, we will start requiring them to provide execution/calldata for AMM interactions that they internalize.

This PR changes the `exec_plan` that is returned from HTTP solvers to instead use an enum to indicate whether or not it is "internal" or to be included in the execution plan. Currently, all we do with "internal" interactions is to skip them when adding them to the solution and logging a message indicating that it was internalized.

Once we start recording settlement competitions in the database, we need to make sure that these internalized AMM interactions get stored somewhere so that they can be verified at a later date. For this, some additional refactoring is necessary in order to make encoding interactions into settlements easier.

One additional point of interest is that we currently don't verify that internalized interactions are indeed over the list of "allowed buffer trading" tokens. I plan on adding this as an follow up PR to not over-complicate this one.

### Test Plan

Added a unit test that verifies that internal interactions don't get encoded.