Use path filter first before method (#280)
Let's say you get the path and the method wrong. The endpoint is the more significant error so you want to be informed about this first with a 404. The method only matters once we have the correct endpoint so you only want to be informed about it in that case and with a 405. This is what this code change does.