Add Blockscout based token owner finding (#275)
Depends on #274 and partially address #125 

Yearn is working with a bunch of exotic tokens (e.g. [x3CRV](https://blockscout.com/xdai/mainnet/token/0x1337bedc9d22ecbe766df105c9623922a27963ec/token-holders#holders)). In order for those tokens to pass our bad token detector, we need to find a token owner from which we can simulate a bunch of transactions and verify the token is not behaving in a weird way (e.g. taking a fee on transfer).

In the past we indexed individual candidates for pool balances based on our LiquiditySources (e.g. Uni v2 pools, the Balancer vault and the Uni v3 factory). For this yearn token specifically, we could index the Curve Gauge, but I assume that for different tokens we need to add more one-off token finders. We have discussed more generalized ways in the past that would rely on external APIs.

This PR adds the Blockscout token holder API as an implementation which seems to work for all tokens on Ethereum and xDAI.

### Test Plan

e2e test for the API and checking that we now can place orders like (even when running with tracing API enabled):

```
curl 'http://localhost:8080/api/v1/quote' \
                                                       -H 'Accept: */*' \
                                                       -H 'Accept-Language: en-GB,en-US;q=0.9,en;q=0.8' \
                                                       -H 'Connection: keep-alive' \
                                                       -H 'Content-Type: application/json' \
                                                       -H 'Origin: https://barn.cowswap.exchange' \
                                                       -H 'Referer: https://barn.cowswap.exchange/' \
                                                       -H 'Sec-Fetch-Dest: empty' \
                                                       -H 'Sec-Fetch-Mode: cors' \
                                                       -H 'Sec-Fetch-Site: cross-site' \
                                                       -H 'Sec-GPC: 1' \
                                                       -H 'User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/101.0.4951.67 Safari/537.36' \
                                                       -H 'X-AppId: 0xE4D1AB10F2C9FFE7BDD23C315B03F18CFF90888D6B2BB5022BACD46AB9CDDF24' \
                                                       --data-raw '{"kind":"sell","sellAmountBeforeFee":"10000000000000000000","sellToken":"0xe91D153E0b41518A2Ce8Dd3D7944Fa863463a97d","buyToken":"0x1337BedC9D22ecbe766dF105c9623922A27963EC","from":"0xF062Fe954952012EB7e587C6f90A4549DE717dcA","receiver":"0xF062Fe954952012EB7e587C6f90A4549DE717dcA","appData":"0xE4D1AB10F2C9FFE7BDD23C315B03F18CFF90888D6B2BB5022BACD46AB9CDDF24","validTo":1655464161,"partiallyFillable":false}' \
                                                       --compressed
```
### Release notes

We should keep an eye on this added external dependency (although in case of failure it should just be silently ignored).