Remove fallback route (#279)
We have a warp fallback route based on `warp::any()` that facilitates collecting timing metrics even of requests that get rejected. The problem with this approach is that it swallows all previous rejections so our `.recover(handle_rejection)` is never called because there are no rejections. It also makes the error returned in the http response NOT_FOUND instead of a more specific error like "invalid method" or "request body too large".
I am getting rid of the fallback because timing information on rejections is less important than better error messages. Rejections are all cheap to process so we will not need their timing information. With this change we still keep track of the number of rejections in a metric that previously was never set as the code wasn't reached.