Initial OrderQuoting implementation (#284)
Follow up to #283

This PR provides a `OrderQuoter2` which implements the newly introduced `OrderQuoting` trait introduced in #283.

The logic in here is a combination of the logic from `OrderQuoter::calculate_fee_parameters` and `MinFeeCalculator::compute_unsubsidized_min_fee`. What is nice about combining these two methods is that we can reuse the price estimate (although - in practice, this was already reused in the important cases because use `RequestSharing<>` to reuse in-flight requests to external APIs).

Note that this order quoter is still not being used. Additionally it does **not** apply subsidies to the fee amount. I want to refactor how we do fee subsidies before adding it to the new `OrderQuoting` implementation and didn't want to make this PR too big in the process.

Implementation for `QuoteStoring` backed by our Postgres database is also missing and will be introduced before we start making use of this quoter.

### Test Plan

CI. Added new test cases for the introduced logic.